requirements.txt
Flask
Flask-Cors
google-cloud-aiplatform==1.46.0 
langchain==0.1.14
langchain_google_vertexai==0.1.2
chromadb==0.4.24
unstructured==0.12.6
pillow-heif==0.15.0
unstructured-inference==0.7.25
pypdf==4.1.0
pdf2image==1.17.0
unstructured_pytesseract==0.3.12
pikepdf==8.14.0
pdfminer.six==20221105
faiss-cpu